// src/types.ts
export interface ProgressEntry {
  date: string;
  weight: number;
  bmi: number;
}

export interface SleepEntry {
  date: string; // YYYY-MM-DD
  hours: number;
}

export interface WaterEntry {
  date: string; // YYYY-MM-DD
  glasses: number;
}

export interface BodyCompositionEntry {
    date: string; // YYYY-MM-DD
    bmi: number;
    bodyFatPercentage: number;
    leanBodyMass: number;
    muscleMass: number;
    visceralFatIndex: number;
    proteinMass: number;
    proteinPercentage: number;
}

export interface FastingEntry {
  date: string; // YYYY-MM-DD
  startTime: string; // e.g., "10:00 AM"
  endTime: string; // e.g., "6:00 PM"
  duration: number; // in hours
}

export interface WorkoutLogEntry {
  date: string; // YYYY-MM-DD
  name: string; // e.g., "General Fat Loss Workout"
  duration: number; // in minutes
}

export interface LoggedMeal {
  name: string;
  calories: number;
}

export interface DailyIntake {
  date: string; // YYYY-MM-DD
  loggedMeals: LoggedMeal[];
  otherCalories: number;
  totalIntake: number;
  targetCalories: number;
}